+++ CHANGELOG.md
@@ -29,6 +29,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `NablaException.Authentication.NotAuthenticated` will now be correctly sent as an error in the `Flow`
   for methods that require authentication and return a `Flow`, rather than thrown.
 - Fix usage of color selector as background not working on older android api levels.
+- Fix media picker bottom sheet not fully shown when on landscape.
 
 ## [1.0-alpha03] - 2022-05-16